pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=64", "wheel", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "protinfo"
dynamic = ["version"]
description = "MCCE tool for obtaining structural info about a protein & how MCCE processes it in step1."
readme = "README.md"
authors = [
    {name = "Cat Chenal @ GunnerLab"},
]
maintainers = [
    {name = "Cat Chenal @ GunnerLab"},
]
license = {file = "LICENSE"}
keywords = [
    "biophysics",
    "proteomics",
    "cheminformatics",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Biophysicists, Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
 ]

requires-python = ">=3.10"
dependencies = [
    "biopython",
    "Jinja2",
    "pandas",
    "ipykernel",
    "pytest",
]

[project.urls]
Source = "https://github.com/GunnerLab/MCCE_ProtInfo"

[tool.setuptools.packages]
find = {}

[project.scripts]
ProtInfo = "protinfo.cli:prot_info_cli"

[tool.setuptools_scm]
version_file = "protinfo/_version.py"
